# app/src/main/resources/static/openapi.yaml
# this is an example of the Uber API
# as a demonstration of an API spec in YAML
swagger: "2.0"
info:
  title: URL SHORTENER
  description: A url shortener is a tool to reduce a long link and then share it.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
# array of all schemes that your API supports
schemes:
  - http
# will be prefixed to all paths
basePath: /
consumes: 
  - application/json
  - application/x-www-form-urlencoded
produces:
 - application/json
paths:
  /api/link:
    post:
      operationId: postlink
      summary: create a shortened URL
      description: Allows a user to create a shortened URL
      parameters:
        - name: request
          in: body
          description: The uri to short
          schema:
            type: object
            required:
              - url
            properties:
              url:
                type: string
                example: https://www.google.com
              limit:
                type: integer
                example: 4
              lat:
                type: number
                example: 41.5421788
              lon:
                type: number
                example: -0.991461
              qr:
                type: boolean
                example: true

      tags: 
        - Uniform Resource Identifier
      responses:  
        "201":
          description: Application is successfully processed
          headers: 
            Location:
              type: string 
              description: Absolute path to the created trimmed URI
          schema:
            type: object
            required:
              - url
            properties:
              url:
                type: string
                description: "URL to short begining with [\"http://www.\"]"
                example: https://www.google.com
              limit:
                type: integer
                description: "Maximum number of redirections of the ShortURL"
                example: 4
              lat:
                type: number
                description: "Latitude of the user that created the shortURL"
                example: 41.5421788
              lon:
                type: number
                description: "Latitude of the user that created the shortURL"
                example: -0.991461
              qr:
                type: boolean
                description: "Whether a QR has to be created or not"
                example: true
              properties:
                type: json
                example: {"error": "URI de destino no validada todavía"}
        "400":
          description: URI cannot be created
          schema: 
            type: object
            $ref: '#/definitions/Error400'
      x-codeSamples:
        - lang: 'Curl'
          source: |
            curl "http://localhost:8080/api/link" \
              -X POST \
              -H "Content-Type: application/json" \
              -d '{"url":"http://www.google.com","limit": 4,"lat":41.5421788,"lon":-0.991461,"qr":true}'
        - lang: 'JavaScript'
          source: |
            uri = "http://localhost:8080/api/link"

            headers = new Headers()
            headers.append("Content-Type", "application/json");
            
            payload = {
              "url": "http://www.google.com",
              "limit": 4,
              "lat": 41.5421788,
              "lon": -0.991461,
              "qr": true
            }
            
            fetch(uri, {
              method: "POST",
              headers,
              body: JSON.stringify(payload)
            }).then(res => res.json())
        - lang: 'Python'
          source: |
            import requests

            uri = "http://localhost:8080/api/link"
            
            headers = {"Content-Type": "application/json"}
            
            payload = {
              "url": "http://www.google.com",
              "limit": 4,
              "lat": 41.5421788,
              "lon": -0.991461,
              "qr": True
            }
            
            response = requests.post(uri, headers=headers, json=payload)
  /api/link/{id}:
    get:
      operationId: getlink
      summary: Information about the URI
      description: Returns relevant information about the shortened URI identified by the parameter id
      parameters:
        - name: id
          in: path
          description: Identifier of the URI
          required: true
          type: string
      tags: 
        - Uniform Resource Identifier
      responses:  
        "200":
          description: Trimmed URI exists and it has been confirmed that it can be used for redirects
          schema:
            type: object
            $ref: '#/definitions/shorturl'
        "400":
          description: Trimmed URI exists and it has not yet been confirmed whether or not the redirection can be made
          headers: 
            Retry-After:
              type: number 
              description: Time in second to wait until next request
          schema: 
            type: object
            $ref: '#/definitions/Error400-not-validated'
        "403":
          description: Trimmed URI exists and cannot be used for redirections
          schema: 
            type: object
            $ref: '#/definitions/Error403'
        "429":
          description: Shortened URI exists and too many requests have been sent during a given amount of time
          headers:
            Retry-After: 
              type: number
              description: Time in second to wait until next request
          schema: 
            type: object
            $ref: '#/definitions/Error429'
      x-codeSamples:
        - lang: 'Curl'
          source: |
            curl "http://localhost:8080/api/link/4170157c"
        - lang: 'JavaScript'
          source: |
            id = "4170157c"
            uri = `http://localhost:8080/api/link/${id}`
            
            fetch(uri).then(res => res.json())
        - lang: 'Python'
          source: |
            import requests

            id = "4170157c"
            uri = f"http://localhost:8080/api/link/{id}"
            
            response = requests.get(uri)
  /{id}:
    get:
      operationId: redirect
      summary: Get redirect
      description: Performs redirection to a URI identified by the parameter id
      parameters:
        - name: id
          in: path
          description: Identifier of the URI
          required: true
          type: string
      tags: 
        - Uniform Resource Identifier
      responses:
        "307":
          description: Trimmed URI exists and it has been confirmed that the redirection can be performed
          headers:
            Location: 
              type: string
              description: Absolute path to the created trimmed URI
        "400":
          description: Trimmed URI exists and it has not yet been confirmed whether or not the redirection can be made
          headers: 
            Retry-After:
              type: string 
              description: time in seconds until next request
          schema: 
            type: object
            $ref: '#/definitions/Error'
        "403":
          description: Trimmed URI exists and cannot be used for redirections
          schema: 
            type: object
            $ref: '#/definitions/Error'
        "429":
          description: Shortened URI exists and too many requests have been sent during a given amount of time
          headers:
            Retry-After: 
              type: number
              description: Time in second to wait until next request
          schema: 
            type: object
            $ref: '#/definitions/Error'
      x-codeSamples:
        - lang: 'Curl'
          source: |
            curl "http://localhost:8080/4170157c"
        - lang: 'JavaScript'
          source: |
            id = '4170157c'
            uri = `http://localhost:8080/${id}`
            
            fetch(uri).then(res => res.json())
        - lang: 'Python'
          source: |
            import requests

            id = '4170157c'
            uri = f"http://localhost:8080/{id}"
            
            response = requests.get(uri)
  /{id}/qr:
    get:
      operationId: qr
      summary: Get QR
      description: Returns a QR code containing the absolute path of the GET /{id} request.
      parameters:
        - name: id
          in: path
          description: Identifier of the URI
          required: true
          type: string
      tags: 
        - Uniform Resource Identifier
      responses:
        "200":
          description: Trimmed URI exists and it has been confirmed that the redirection can be performed
          headers: 
            Content-Type:
              type: string
              enum: 
                - image/png
              description: The format used by the QR code is PNG
          schema:
            type: object
            $ref: '#/definitions/Error' 
        "400":
          description: Trimmed URI exists and it has not yet been confirmed whether or not the redirection can be made
          headers: 
            Retry-After:
              type: string 
              description: Time in seconds until next request
          schema: 
            type: object
            $ref: '#/definitions/Error'
        "403":
          description: Trimmed URI exists and cannot be used for redirections
          schema: 
            type: object
            $ref: '#/definitions/Error'
        "429":
          description: Shortened URI exists and too many requests have been sent during a given amount of time
          headers:
            Retry-After: 
              type: number
              description: time in second to wait until next request
          schema: 
            type: object
            $ref: '#/definitions/Error'
      x-codeSamples:
        - lang: 'Curl'
          source: |
            curl "http://localhost:8080/4170157c/qr"
        - lang: 'JavaScript'
          source: |
            id = '4170157c'
            uri = `http://localhost:8080/${id}/qr`
            
            fetch(uri).then(res => res.json())
        - lang: 'Python'
          source: |
            import requests

            id = '4170157c'
            uri = f"http://localhost:8080/{id}/qr"
            
            response = requests.get(uri)
definitions:
  shorturl:
    properties:
      url:
        type: string
        description: "Shortened URL"
        example: "http://localhost:8080/4170157c"
      target:
        type: string
        description: "Target URL"
        example: "http://www.google.com"
      lat:
        type: number
        description: "Latitude of the user that created the shortURL"
        example: 41.5421788
        x-nullable: true
      lon:
        type: number
        description: "Longitude of the user that created the shortURL"
        example: -0.991461
        x-nullable: true
      qr:
        type: string
        description: "Link to the QR endpoint"
        example: "http://localhost:8080/4170157c/qr"
        x-nullable: true
      safe:
        type: boolean
        description: "Whether the uri is safe or not"
        example: true
  Error:
    properties:
      message:
        type: string
  Error400:
    properties:
      statusCode:
        type: integer
        description: "Http status"
        example: "400"
      message:
        type: string
        description: "Description of the error"
        example: "[https://ww.google.com] does not follow a supported schema"
      timestamp:
        type: string
        description: "Timestamp of the request"
        example: "2022-12-19T23:58:57.70665+01:00"
  Error400-not-validated:
    properties:
      statusCode:
        type: integer
        description: "Http status"
        example: "400"
      message:
        type: string
        description: "Description of the error"
        example: "URI has not been validated yet"
      timestamp:
        type: string
        description: "Timestamp of the request"
        example: "2022-12-19T23:58:57.70665+01:00"
  Error403:
    properties:
      statusCode:
        type: integer
        description: "Http status"
        example: "403"
      message:
        type: string
        description: "Description of the error"
        example: "URI is not safe"
      timestamp:
        type: string
        description: "Timestamp of the request"
        example: "2022-12-19T23:58:57.70665+01:00"
  Error429:
    properties:
      statusCode:
        type: integer
        description: "Http status"
        example: "429"
      message:
        type: string
        description: "Description of the error"
        example: "Too many redirections for URL [https://ww.google.com]"
      timestamp:
        type: string
        description: "Timestamp of the request"
        example: "2022-12-19T23:58:57.70665+01:00"
  